requirements.txt
aiogram>=3.17.0
flask
google-generativeai
python-dotenv
pydantic>=2.10.5
aiocryptopay
Pillow